- Add IngHierarchy response structure - Move requests into their own package
