+++ src/main/java/com/ing/inghierarchy/service/PersonService.java
@@ -8,8 +8,8 @@ import com.ing.inghierarchy.repositories.ManagerRepository;
 import com.ing.inghierarchy.repositories.RoleRepository;
 import com.ing.inghierarchy.repositories.TeamMemberRepository;
 import com.ing.inghierarchy.repositories.TeamRepository;
-import com.ing.inghierarchy.web.ManagerRequest;
-import com.ing.inghierarchy.web.TeamMemberRequest;
+import com.ing.inghierarchy.web.request.ManagerRequest;
+import com.ing.inghierarchy.web.request.TeamMemberRequest;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
