+++ src/main/java/com/ing/inghierarchy/web/TeamController.java
@@ -2,6 +2,7 @@ package com.ing.inghierarchy.web;
 
 import com.ing.inghierarchy.domain.Team;
 import com.ing.inghierarchy.service.TeamService;
+import com.ing.inghierarchy.web.request.TeamRequest;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
