+++ src/main/java/com/ing/inghierarchy/web/response/ManagementChain.java
@@ -1,0 +1,16 @@
+package com.ing.inghierarchy.web.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.Accessors;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Accessors(chain = true)
+public class ManagementChain {
+    private ManagerResponse manager;
+}
