+++ src/main/java/com/ing/inghierarchy/web/response/TeamMemberResponse.java
@@ -1,0 +1,20 @@
+package com.ing.inghierarchy.web.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.Accessors;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Accessors(chain = true)
+public class TeamMemberResponse {
+
+    private String id;
+    private String corporateId;
+    private String role;
+    private String name;
+}
