+++ src/test/java/com/ing/inghierarchy/web/PersonControllerTest.java
@@ -6,6 +6,8 @@ import com.ing.inghierarchy.domain.TeamMember;
 import com.ing.inghierarchy.repositories.ManagerRepository;
 import com.ing.inghierarchy.repositories.TeamMemberRepository;
 import com.ing.inghierarchy.service.PersonService;
+import com.ing.inghierarchy.web.request.ManagerRequest;
+import com.ing.inghierarchy.web.request.TeamMemberRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
