+++ src/test/java/com/ing/inghierarchy/web/TeamControllerTest.java
@@ -3,6 +3,7 @@ package com.ing.inghierarchy.web;
 import com.ing.inghierarchy.Exceptions.IngHttpException;
 import com.ing.inghierarchy.domain.Team;
 import com.ing.inghierarchy.service.TeamService;
+import com.ing.inghierarchy.web.request.TeamRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
